created new module in common for conversion functions. Fixed cmake. Cleaned up code
