+++ src/common/interface/converters/converters.hpp
@@ -1,14 +1,15 @@
-#pragma once
+#ifndef CONVERTERS__CONVERTERS_HPP
+#define CONVERTERS__CONVERTERS_HPP
 
 #include <common_ros2/msg/posture.hpp>
 #include <common_ros2/msg/spline.hpp>
-#include <core_datastructures/Point.hpp>
+// #include <core_datastructures/Point.hpp>
 #include <core_datastructures/Posture.hpp>
-#include <core_datastructures/Pose.hpp>
+// #include <core_datastructures/Pose.hpp>
 
-namespace convert{
+namespace converters{
     /**
-     * @brief 
+     * @brief Convert ros2 Posture to domain
      * 
      * @param posture 
      * @return core_datastructures::Posture 
@@ -17,10 +18,12 @@ namespace convert{
 
 
     /**
-     * @brief 
+     * @brief Convert domain Posture to ros2
      * 
-     * @return common_ros2::msg::Possture 
+     * @param posture
+     * @return common_ros2::msg::Posture 
      */
-    common_ros2::msg::Possture to_ros2(const core_datastructures::Posture);
+    common_ros2::msg::Posture to_ros2(const core_datastructures::Posture& posture);
 }
 
+#endif // CONVERTERS__CONVERTERS_HPP
