+++ src/common/interface/package.xml
@@ -8,6 +8,7 @@
   <license>TODO: License declaration</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
+  <buildtool_depend>rosidl_default_generators</buildtool_depend>
 
   <build_depend>common_domain</build_depend>
   <build_depend>rosidl_default_generators</build_depend>
